Fixing open issues
So what did I do?
I checked the JSON that the AJAX is sending. It is OK. So I modified the
WebController to parse the json into a UserOrder object, and then pass
the object to CoffeeOrderHandler (which is now Autowired).
I fixed the MailSenderImpl to add all of the requested recipients, so no
more work should be there basically.
I created new OrdersRepositoryInMemoryImpl to be the Component for
OrdersRepository instead of OrdersRepositoryImpl. I didn't test it, but
it should work fine.

And what is left to do?
Verify that the project can be deployed properly, as I was uncommenting
the  CoffeeOrderHandler class.
I could see in the logs that even that WebController received the order
call, it didn't succeeded to return an OK status to the client, and
that's why we saw an ERROR in the UI. It is critical to find out how to
return the proper response, as the AJAX is accepting. I strongly suggest
to involve Irit for this issue.